requirements.txt
Flask==3.0.0
torch
torchvision
ultralytics
numpy
opencv-python-headless
Pillow
gunicorn==21.2.0
